pyproject.toml
[project]
name = "C14H22O-bot"
version = "0.1.1"
description = "一个没有什么用的QQ官方bot"
authors = [
    {name = "C14H22O", email = "160833462+C14H22O@users.noreply.github.com"},
]
dependencies = [
    "nonebot2[websockets]>=2.2.1",
    "nonebot-adapter-qq>=1.4.2",
    "nonebot-plugin-fishing>=0.2.1",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.nonebot]
plugins = ["nonebot_plugin_fishing"]
